Split vm code into three additional files No code changes done
